SIMPLE_LINEAR REGRESSION MODEL
This is my submission of the task 1 (Prediction using Supervised ML) as a Data Science & Business Analytics Intern under GRIP at The Sparks Foundation. This is a simple linear regression model that involves two variables:: no. of hours a student studies and marks of the student and predicts the percentage of an student based on the no. of study hours.